// src/shared/ui/UIButton/types.ts
// import { ButtonHTMLAttributes } from "vue";

export enum ButtonVariant {
    DEFAULT = 'default',
    CLEAR = 'clear',
    OUTLINE = 'outline',
    BACKGROUND = 'background',
    BACKGROUND_INVERT = 'backgroundInvert',
}
export enum ButtonSize {
    M = 'size_m',
    L = 'size_l',
    XL = 'size_xl',
}
// export interface ButtonProps extends ButtonHTMLAttributes {
//     // className?: string
//     variant?: ButtonVariant
//     squire?: boolean
//     size?: ButtonSize
// }